backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
orjson==3.10.15
numpy==2.1.3
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
rank-bm25==0.2.2
pymupdf==1.24.14
httpx==0.28.1
rapidfuzz==3.11.0
